refactor: features/run/state → context 폴더명 변경 (Phase 2.3)
- state/ 폴더를 context/로 이름 변경
- 11개 파일의 import 경로 업데이트
- 폴더 구조 명확화 (Context API 관련 파일임을 명시)
