+++ src/features/run/components/RunStatsPanel.tsx
@@ -2,7 +2,7 @@ import StatsIndicator from "@/src/components/ui/StatsIndicator";
 import { TextWithSub } from "@/src/components/ui/TextWithSub";
 import { Typography } from "@/src/components/ui/Typography";
 import { Telemetry } from "@/src/apis/types/run";
-import { RunStatus } from "@/src/features/run/state/context";
+import { RunStatus } from "@/src/features/run/context/context";
 import { View } from "react-native";
 
 export interface StatsDisplayItem {
