+++ src/features/run/hooks/useRunningSession.ts
@@ -4,8 +4,8 @@ import { useEffect, useMemo, useReducer, useRef } from "react";
 import "react-native-get-random-values";
 import { v4 as uuidv4 } from "uuid";
 import { useRunMetronome } from "../../audio/useRunMetronome";
-import { RunAction } from "../state/actions";
-import { initialRunContext, runReducer } from "../state/reducer";
+import { RunAction } from "../context/actions";
+import { initialRunContext, runReducer } from "../context/reducer";
 import { joinedState } from "../store/joinedState";
 import { RunMode } from "../types";
 import { CourseMetadata, CourseVariant } from "../types/status";
